+++ packages/moonstone/Slider/Slider.js
@@ -44,13 +44,23 @@ const PrivateSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 		propTypes: /** @lends moonstone/Slider.SliderBase.prototype */{
 
 			/**
+			 * Overrides value string to read for the slider. It is mapping to `aria-valuetext` internally.
+			 * By default, `aria-valuetext` is set to the current value. This should only be used
+			 * when the parent controls the value of the slider directly through the props.
+			 *
+			 * @type {String|Number}
+			 * @public
+			 */
+			'aria-label': PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+
+			/**
 			 * Overrides the `aria-valuetext` for the slider. By default, `aria-valuetext` is set
 			 * to the current value. This should only be used when the parent controls the value of
 			 * the slider directly through the props.
 			 *
 			 * @type {String|Number}
 			 * @memberof moonstone/Slider.SliderBase.prototype
-			 * @public
+			 * @private
 			 */
 			'aria-valuetext': PropTypes.oneOfType([PropTypes.string, PropTypes.number]),